Create DecimalToBinary convert
convert decimal to binary: c language code to convert an integer from decimal
number system(base-10) to binary number system(base-2). Size of integer is assumed 
to be 32 bits. We use bitwise operators to perform the desired task. We right shift 
the original number by 31, 30, 29, ..., 1, 0 bits using a loop and bitwise AND the
number obtained with 1(one), if the result is 1 then that bit is 1 otherwise it is 0(zero).